GHE: libpldmresponder: implement responder for readFileByType
This commit implements the handler for readFileByType. This will
transfer 1-15bytes of data in inband mode for specific file
types, like LID, DUMP etc. The data to be transferred via this command are
either the left overs from the allowed DMA transfer or a small file
not eligible to be DMAed.

Change-Id: I9db77f78b0aa4e289fb0ab940121569b431f4dfe
Signed-off-by: Sampa Misra <sampmisr@in.ibm.com>