PLDM: Hex value support for the PCIE config space data
This commit sets the PCIE config space data sent from
host with a prefix '0x' and the hex value. This change
was needed as the Redfish validator expects the fields
to be in a particular Regex form otherwise it fails to
validate.

Tested:
By using busctl commands to set the fields with the data
sent and also ran the Redfish validator to verify.

Signed-off-by: Pavithra Barithaya <pavithra.b@ibm.com>