Support for Mex IO Enclosure version
PHYP sends a FRU record set PDR and send the io enclosure
version data via the frurecord table, and pldm parses that
and hosts the Version Interface on the mex chassis object.

Currently phyp does not have the support for this yet, so
hardcoded a dummy value for now, would have to remove that part
of code once the phyp change is in the driver.

Signed-off-by: Manojkiran Eda <manojkiran.eda@gmail.com>