Host availability interface on cores & mex dbus objects
When the host is off, the dbus objects for cores and mex
would be just dummy and a user would be not able to do any
operation on those.

So, PLDM would host the Availability interface on all the
dbus object and set the availability to false when the host
is off, and set them back to true right after the PDR exchange.

There should be a corresponding change required in redfish
to look for the availbility state and mark the redfish state
as Unavailable offline.

Signed-off-by: Manojkiran Eda <manojkiran.eda@gmail.com>