PLDM: fix for SW543085
There was a bug in PLDM were post reboot when the host is up
we talk to HBRT to get sensor states of OCC and cores etc.
and when we power off that is not reverted back to teh valid eid.
This was not triggered till now as we were getting bad TL PDR for
HBRT.

Signed-off-by: Pavithra Barithaya <pavithra.b@ibm.com>