oem_ibm: Use version path for persistent version object
The version id is now a random number in order to support the upload of
the same version:
[1] https://gerrit.openbmc-project.xyz/c/openbmc/phosphor-bmc-code-mgmt/+/49972

Therefore the version id would be different when uploaded and after a
reboot. A unique identifier has been added to the Path property, this
specifies the location on flash of the version, on emmc is the partition
"a" or "b". Therefore, use that property instead of the version id to
know if the bmc has booted from the same version or not.

Tested:
- After code update:
Feb 11 19:42:04 rain118bmc pldmd[1545]: BMC have booted with the new
image runningPath=a
Feb 11 19:42:04 rain118bmc pldmd[1545]: Previous Image was: b

root@p10bmc:~# cat /var/lib/pldm/bootSide
Temp
Temp
a

- After reboot on the same image:
Feb 11 19:49:56 rain118bmc pldmd[601]: BMC have booted with the previous
image runningPath=a

Change-Id: I8b9713a7992525336169f3f639b4ef9ea77392d6
Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>