oem: ibm: Set next boot side attribute
During a rename, set the next boot side to Perm (same as current) so
that in case of a failed update, PHYP knows to boot from the Perm side.
The next boot side would be set to Temp on a successful update (via the
Priority callback function).

Update the next boot side attribute (fw_boot_side) whenever the next
boot side value is set or changed so that in error scenarios, PHYP knows
which side to boot from.

Setting fw_boot_side attribute automatically calls the function to set
the next boot side because the attribute can be also be set externally
by PHYP, therefore check the requested next boot side already has the
highest priority, in which case there's no need to set the Priority.
Also checking the current priority allows to catch if the alternate
version may not exist because of a code update that failed or is in
progress. Once an image is uploaded, set the non running version to the
uploaded image since the current non running image will be deleted
during the update.

Tested: Verified the fw_boot_side attribute matched the information in
the persisted bootSide file. The Priority value was only set when needed
(the requested next boot side was different than the current next boot
side). During a rename, both current and next are set to Perm.

Change-Id: Ibf44f53769a3394826796b1d101fcc39bdd2a4ef
Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>