oem_ibm: Fix for dump offload and system dump delete issues
This commit is to fix the dump offload issue by using
GetManagedObjects instead of GetSubTreePaths to search for the
dump entry match. This also adds the support the delete of
non-disruptive system dump entry in non-hmc managed scenario
after a successful offload of the dump to OS.

Tested by triggering a non-disruptive system dump using an empty
vsp string. Also tested the system dump and resource dump offloads.

CQ: SW544034

Signed-off-by: Jayashankar Padath <jayashankar.padath@in.ibm.com>
Change-Id: Ia46c2f3b654d509c6669e909ff803958f5a8f4ea