oem: ibm: Close fd after transferFileData
The file descriptor was being closed before the call to transferFileData
which performs a write to it, move the close call after the call to
write. Also add an fsync call to ensure the data has made it to the
filesystem and can be processed.

Tested: Performed an inband update.

Change-Id: I199782e2a1b96496d57f3f6e840a298ba66de9c5
Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>