Remove Default Option
Default should not be an option for the current eBMC system. This
option was applicable for FSP and ASMI based systems. Rainier and
Everest based systems should not have this option. The default
choice needs to be removed from pvm_default_os_type and
pvm_default_os_type_current, and the default value needs to be
changed to AIX.

Defect SW547992 :
https://w3.rchland.ibm.com/projects/SLIC/bestquest/index.php?defect=SW547992