oem-ibm:Unmap the memory selectively
These changes will handle the case where host crashes in the
middle of outofband DMA transfer.

With the existing code, after the DMA transfer PLDM unmaps
the memory and closes the file handler(XDMA FD).
Closing the XDMA file handler will reset the XDMA Driver which
internally will unmap the memory if not already unmapped.

But when the host crashes during the DMA transfer,
PLDM receives EINTR(interrupt) and it unmaps the memory.
XDMA driver doesn't know that the operation is interrupted,
it still thinks that operation is in progress.
munmap will hang if xdma driver thinks the operation is in progress.

To solve the PLDM Hang in munmap, in case of EINTR during DMA transfer
PLDM should skip munmap and just close the file handler.
Once PLDM closes the file handler driver gets reset and
it munmaps the memory which is not unmapped yet.

Tested: Powered on the host and confirmed if the munmap is called
for all possitive cases.
Hardcoded return value to EINTR and confirmed that memory
is not getting unmapped

Signed-off-by: ArchanaKakani <archana.kakani@ibm.com>