oem: ibm: Do not remove the staging directory
The staging directory is created by the software manager application,
therefore it should not be deleted by pldm or it won't be re-created
until the next BMC reboot, and the inband update would fail with:

Apr 29 20:21:53 p10bmc pldmd[1980]: Cannot stat source directory
"/var/lib/phosphor-software-manager/hostfw/staging/lid" because No such
file or directory

Instead iterate over the contents of the directory to clear them.

Tested: The contents of the staging directory got removed, but not the
staging directory itself.

Change-Id: Id76ac325d49ca48891c99e7b71b1a0d5d64c85c3
Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>