[OEM-IBM]Send TM keyword to PHYP reading FC keyword from Dbus
PLDM sends the VCEN records(both FC and TM keywords) to phyp via the
bmc fru record table. But TM is not programmed in VPD and tools like
inventoryscout on AIX looks at TM keyword to parse inventory data.

As reprogramming this in VPD during manufacturing is not feasible,
pldm will do a work around to read FC key word from the dbus and
populate the TM Fru entry in the fru record-table.

Updated Json files to pick FC keyword from Dbus and type remains
the same as that of TM.

Fixes: SW550725

Tested using pldmtool

Signed-off-by: ArchanaKakani <archana.kakani@ibm.com>