fix to persist the indicator status after BMC R/R (#235)
This commit sets the present and the event state of the sensors
same. when the led was turned on before R/R the sensor state is updated.
After R/R event state of the sensor contains the changed state but the
present state was pointing to old state. the host was looking at the
present state field of the sensor after the R/R.
- adds the setOperationStatus in createDbusObjects method

fixes - SW550014

Signed-off-by: Pavithra Barithaya <pavithra.b@ibm.com>