oem: ibm: Move the NVRAM files to the hostfw directory
Move the NVRAM files from /var/ on the BMC to the hostfw directory so
that they don't get erased during a BMC factory reset.

To help migrate existing systems, check if the files exist in /var/, and
if they do, move them to the hostfw destination.

Tested: Verified that the PHYP NVRAM files were moved if they already
existed in /var/, and that empty files were created in hostfw if they
didn't.

Change-Id: I8d30d38e10f6f699569fc437210331bacabfc641
Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>