PLDM: Reset image activation after sending end update event
ImageActivation match which is reset even before we receive
the Image Active signal is received. Which is why we are
failing to send the End update sensor event to PHYP. This
needs to be reset only after the end update success/failure
event to PHYP after receiving the Active signal on the activation
property.

Defect:SW551662

Signed-off-by: Sagar Srinivas <sagar.srinivas@ibm.com>