fixes MEX CM and looping issue in MEX systems
In MEX systems, PLDM receives multiple refresh repo events from host.
After receiving refresh repo event, PLDM delete the host PDRs except the
Entity association PDRs and then fetches list of PDRs from host again.
In this scenario PDRs are added such a way that it causes loop.

Fixed the issue by merging Entity association PDRs(from Phyp) in the BMC range.
Following are the modification:

    Using record handle 0xFFFFFFFF to mention that PDR has to be added to
    BMC range
    Modified oem function to traverse the complete repository and get the
    latest BMC PDR

For MEX CM - whenever connectors are present under the card and a
CM is done on the card, it was not handled properly and PHYP task
would crash. With this changes it is fixed.

Tested:
Tested using pldmtool and busctl command

    pldmtool platform getpdr -a
    Always returned expected number of PDRs without causing a loop
    busctl tree PLDMD
    PLDM tree had all the objects populated
    Tried Power on/off operation and verified pdr list
    Same with reset reload operation

fixes : SW551128

Signed-off-by: Pavithra Barithaya <pavithra.b@ibm.com>