oem-ibm dump offload :Unmap the memory selectively
These changes will handle the case where host crashes in the
middle of outofband DMA transfer during system dump offload.

When the host crashes during the DMA transfer,
PLDM receives EINTR(interrupt) and it unmaps the memory.
XDMA driver doesn't know that the operation is interrupted,
it still thinks that operation is in progress.
munmap will hang if xdma driver thinks the operation is in progress.