ensure proper dependency in phyp nvram services (#292)
Some changes were recently introduced that affected the startup order of
services on the way to bmc ready state. This change brought to light a
missing dependency in these service files. They depend on the nvram
links being setup correctly in var/lib/phosphor-software-manager/hostfw/
so therefore need to be sure the service that creates that link has run.

The service that creates the /media/hostfw/nvram dir is also needed.

The change that added this dependency is not in upstream so this fix is
in downstream only.

Signed-off-by: Andrew Geissler <geissonator@yahoo.com>