Fix topology links for flett and pcie switch
pldm is confused between the flett links that has the flett
involved (io_slot location codes are cec slots) & the links
that involve pcie switch on the cable card(io_slot location
codes are mex slots).

So the fix is to first check if the local ports are on the
cec first, and if not, then check for mex so that we populate
both the flett and the links that capture the pcie switch on the
cable card.

Signed-off-by: Manojkiran Eda <manojkiran.eda@gmail.com>