PLDM: Changes to persist the HB PDRs in the repo before merging
This commit persists the HB range PDRs in the repo before
they are merged. This change was needed as PHYP would not send
down BMC, the entity Association PDRs of HB after a R/R as the
PDRs after a merge will be in BMC range of record handles.
Due to this the pldm ebmcModel in PHYP was not updated properly.

This commit also prevents adding duplicate terminus locator
PDRs to the repo.

Tested: Multiple poweron and  poweroff
        Multiple Reset Reloads

Fixes: SW553100

Signed-off-by: Pavithra Barithaya <pavithra.b@ibm.com>