Fix support for link with Unknown Type
In the current state pldm would just drop the link that
has the link type as Unknown, the intent behind this PR
would be to add in necessary support in pldm so that we
store the last link type of a link and use for setting
the topology information.

Tested By:
1. Power on the host, make sure we are at Standby.
2. Get topology information and make sure all the links
   are populated correctly.
3. Inject the error on a link using phyp macros 5 times untill the
link is dead
xmwritereg -mmio -drc 20000010 -offset C00 -length 8 -asb -data 8000000000000000
xmtest -slotErrorRecovery 21010010 19
4. Refresh the topology page and make sure the link is
populated and with accurate information.

Signed-off-by: Manojkiran Eda <manojkiran.eda@gmail.com>