Discard persisted attribute value
Discard the persisted attribute value if the bios attribute
value doesn't fall in the range provided by the current firmware level.
In current implementation we restore the persisted attribute
value without validation for string and integer attributes and
in case of enum attributes we just throw error and don't add the
attribute in the table if it fails.
This causes multiple unrecoverable errors getting logged during
IPL if there is change in bios attribute details between old and new
firware levels.

To fix the issue added changes to validate the persisted value
before restoring it during power on.
Added the check for Integer, String and Enum attributes.
Setting it to the default value if the validation of persisted value fails.

Fixes PE00DYMM

Tested:
Below are the tests performed
1. Normal Power on
2. Reset Reload
3. Modify Integer/String/Enum BIOS Attribute Value and modify
   attribute boundries such that attribute value falls out of
   range and then IPL the system and confirmed that default value is
   set and no unrecoverable error SRC logged

Signed-off-by: ArchanaKakani <archana.kakani@ibm.com>