fix to handle PHYP hang after a NEX drawer CM
This bug was introduced due to SW553100 changes where we started
adding correct terminus handles to entity association PDRs in our repo.
Due to this when we get a refreshEntireRepository repo change event
from PHYP, BMC removes all the PHYP PDRs including the entity association
PDRs. The BMC entity tree is untouched.

When a MEX drawer is powered on back, our code just checks the BMC entity
tree for the entity and if its present we skip adding the PDR to the repo.
This leads to BMC reusing some of the PDR record handles and PHYP asserts
as it doesn't handle it.

Unit tested below scenarios:
        - PowerOn and PowerOff
        - ResetReload
        - CEC Hotplug and CEC CM
        - MEX CM

Fixes: SW558361

Signed-off-by: Pavithra Barithaya <pavithra.b@ibm.com>