oem: ibm: fileTable: Update permissions
Fix the permissions per specified in the spec:
pldm file_traits:
[2]: Preserve across firmware updates
[1]: ReadWrite
[0]: ReadOnly

The Preserved files are assumed to be ReadWrite so no need to set some
of them to 6, just 4. Whenever pldm implements the check for ReadOnly,
it can check the first byte (value 1) to prevent writes to the ReadOnly
files.
Also remove two entries that are no longer valid.

Tested: Power on a p10bmc system.

Change-Id: I9ca3411644f210e5715ac20b34287b0a993dcf40
Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>