bios_utils: BIOSTableIter: Failure to create iter throws exception
It's possible that the heap allocation invoked by
pldm_bios_table_iter_create() fails. The existing behaviour is to
assert() in the implementation that it hadn't failed, but that will be
changed to instead have the API return NULL.

Check for the NULL condition with the understanding that it may occur in
the future, and if observed, throw an exception to uphold the existing
assumptions in the code.

Signed-off-by: Andrew Jeffery <andrew@aj.id.au>
Change-Id: I20016e8db4c814017306cc68a536a9c232c74e0e
(cherry picked from commit 47266d6fc746904e3ca75b5710bce48c8d06d792)