Return proper Success/Error code to host
This commit adds change to send the right error code
to host during the marker lid validation. When the marker
lid validation fails we were trying to send a PLDM_ERROR
But it was sent as a response to the file write request
and PHYP assumed that the error code returned was for the
write operation and did not go ahead to check the sensor
event sent to indicate the entitlment fail during marker
lid validation.

Defect: PE00LNC2

Signed-off-by: Sagar Srinivas <sagar.srinivas@ibm.com>