PLDM: LG2 migration for missed out traces
As part of adding new features there were
traces that are still using standard ostream
pattern of logging. As PLDM repository migrated
to structured logging hence migrating the new
log traces to follow the same practice.

Additionally minor refinement of traces is also
handled as a part of this commit, listed below -

1. Correction in type of log (ex- error or info)
2. Correction in spelling of log message
3. Removing logs from test file.

Signed-off-by: Riya Dixit <riyadixitagra@gmail.com>