oem-ibm: Change default OS type from "AIX" to "Default" (#493)
This change makes default OS type as "Default" instead of "AIX". Keeping default OS type as AIX can cause issues in systems where AIX is not installed. So the safer option is to keep default OS type as "Default" so PHYP can decide which OS has to boot. With Default, PHYP uses the RB keyword to boot to the default OS.

Tested By : Basic IPL testing and checked the default option on GUI.

Signed-off-by: vkaverap@in.ibm.com <vkaverap@in.ibm.com>
Co-authored-by: vkaverap@in.ibm.com <vkaverap@in.ibm.com>