Review comments & allow password change as needed
This addresses review comments to fix names of the enumeration
constants to add underscores for readability.
This fixes a bug in changePassword.  Before, the changePassword function would
only allow the password to be changed if the password did not need to be changed.
Now it allows the changePassword operation to proceed.

Tested:
Unit test notes: Use the command `journalctl --follow` to see responses from the
various PAM modules as testing proceeds.

Additional step for existing test:
After the password has been set, but before the user is made an admin, ensure
password change is not allowed:
try-pfw-sms-menu test $current_password x$current_password

New test, with setup as before:
passwd --expire test
try-pfw-sms-menu test $current_password x$current_password
Verify the password change was successful, and the password was changed
try-pfw-sms-menu test x$current_password
Verify the authentication was successful

Signed-off-by: Joseph Reynolds <joseph-reynolds@charter.net>