Using pldm_entity as key for object storage
"In order to facilitate the maintenance of 'pldm_entity' within the
'pldm' scope, we have implemented a modification that involves
replacing the use of 'pldm_entity_node' solely during the creation of
the D-Bus map for storing entity D-Bus paths and 'pldm_entity'.
Previously, 'pldm_entity_node' was employed for this purpose, but its
utilization posed challenges within the PLDM due to its opaque nature
as a pointer. Since PLDM cannot directly utilize structure values,
'pldm_entity_node' became less practical for entity maintenance.
Additionally, initializing the 'pldm_entity_node' structure external to
'libpldm' was not feasible. Consequently, we have decided to abandon
the previous commit [1] wherein 'pldm_entity_node' was defined within
'libpldm' with restricted scope. It is imperative to ensure that the
purpose of 'pldm_entity_node' is not violated, as indicated by
reference [2]."

[1] https://gerrit.openbmc.org/c/openbmc/libpldm/+/70340
[2] https://discord.com/channels/775381525260664832/1227797894178803782/1227832985353588797

Testing:
Unit tests passed

Change-Id: I223dee8d5532cc8dfe2259107c804e1a006b5773
Signed-off-by: Kamalkumar Patel <kamalkumar.patel@ibm.com>