hotplug: Add API to check FRU presence
API that checks if FRU with a given object path is present
in the inventory interface is introduced. FRU records are
built only for FRUs that are present in the inventory
interface list. The presence of a FRU is determined by the
"Present" Dbus property.

Tested By: IPL and reset on BMC.

Signed-off-by: Varsha Kaverappa <vkaverap@in.ibm.com>