oem-ibm: Changes to handle the new way of Dump entry
The dump entries are numbered in a new format (for ex:
System dump entry is numbered - A000<>
Resource dump entry is numbered - B000<>
Made changes in code to incorporate these changes.
The code also changes the Token handling and UserChallenge
changes in 1110.

tested: Token and dump flow was successfully tested.

Change-Id: I0a7ddc71af87224d79ba7576b3c94fa8b25e2085
Signed-off-by: Pavithra Barithaya <pavithrabarithaya07@gmail.com>