oem-ibm: Fix in pointer increment in the PCIe Config Data
The PCIe Config data is a 64 bytes huge data which needs to be
parsed and corresponding DBus values needs to be updated. The
parsing logic had a pointer increment issue that resulted in
some bug w.r.t the values.
Removed a debug trace as it is not having much value.

tested on a rainier and blueridge systems.

Signed-off-by: Pavithra Barithaya <pavithrabarithaya07@gmail.com>