Fixing race condition during file transfer
there is a race condition occurs between performing post operation and
sending response to Host. When host sends pels then bmc takes some time
to store the pels and meanwhile host is timed-out so it will try again
to transfer the pels and suddenly BMC finishing the storing and sending
failure status to host which causing the problem.

Signed-off-by: Kamalkumar Patel <kamalkumar.patel@ibm.com>