oem-ampere: eventManager: Handle PCIe hotplug sensor event
This commit adds the handler and needed APIs to handle PCIe hot-plug
event as a Numeric Sensor Event. The handler will decode the event data,
parse them to readable info and log to Redfish Event Log.

Tested:
Hot-plug in/out the PCIe device and see the events logged into Redfish.

Change-Id: I8cc9b712a6f7d9b3402bf72bb2d825ed8d5a8009
Signed-off-by: Chau Ly <chaul@amperecomputing.com>