oem-ibm: Handle the json parse exception correctly
This commit handles the CoD license json parsing exception by
correctly catching it and returns the error to the caller. Also
made the change to avoid the removal of license json file whenever
shared by the host.

Tested:
Done the sanity testing to verify that json file was kept without
removal.

Signed-off-by: Jayashankar Padath <jayashankar.padath@in.ibm.com>